src/requirements.txt
numpy
pandas
scikit-learn
torch
torchvision
matplotlib
seaborn
grpcio
opencv-python
pillow
jupyter
grpcio-tools